TDM-1117: Remove ChromeVox scripts to resolve veracode vulnerabilities
Both of the vulnerabilities were in ChromeVox scripts that this plugin injects
when it determines that the webview the app runs in doesn't already have ChromeVox
(i.e when offline). Prior to android lollipop, webviews with Javascript enabled required an
internet connection for spoken accessibility to work correctly. This is because the webview
relied on a JS-based screen reader (ChromeVox) that may not always be available in the
webview cache. So our solution was to remove these scripts that this plugin injects so
veracode will stop complaining. The only downside is that accessibility users running less
than android lollipop won't be able to get their screen reader working unless their device
is connected to the internet, so they won't hear our "network unavailable" message